+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Version 0.0.17 (upcoming)
 - get_df now also accepts filled_share as kind
+- added function to count the holes in the timeseries depending on there length
 
 ## Version 0.0.16
 - repaired the update_raw function of StationND